nhan khau 1 commit

+++ doanphanmem/src/main/java/com/vanthuandev/doanphanmem/repository/NhanKhauRepository.java
@@ -3,5 +3,5 @@ package com.vanthuandev.doanphanmem.repository;
 import com.vanthuandev.doanphanmem.pojos.NhanKhau;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NhanKhauRepository extends JpaRepository<NhanKhau, Long> {
+public interface NhanKhauRepository extends JpaRepository<NhanKhau, Integer> {
 }
